+++ eliot/tests/test_testing.py
@@ -179,15 +179,21 @@ class LoggedActionTests(TestCase):
         # index 0
         with startAction(logger, "test"):
             # index 1:
-            with startAction(logger, "test"):
+            with startAction(logger, "test2"):
                 # index 2
-                Message.new(x=2).write(logger)
+                Message.new(message_type="end", x=2).write(logger)
             # index 3 - end action
-            # index 4 - end action
+            with startAction(logger, "test3"):
+                # index 4
+                pass
+            # index 5 - end action
+        # index 6 - end action
         logged = self.fromMessagesIndex(logger.messages, 0)
 
         self.assertEqual(
             logged.children[0], self.fromMessagesIndex(logger.messages, 1))
+        self.assertEqual(logged.type_tree(),
+                         {"test": [{"test2": ["end"]}, {"test3": []}]})
 
     def test_ofType(self):
         """
@@ -213,6 +219,10 @@ class LoggedActionTests(TestCase):
             logged, [
                 self.fromMessagesIndex(logger.messages, 1),
                 self.fromMessagesIndex(logger.messages, 5)])
+
+        # String-variant of ofType:
+        logged2 = LoggedAction.ofType(logger.messages, "myaction")
+        self.assertEqual(logged, logged2)
 
     def test_ofTypeNotFound(self):
         """
@@ -303,6 +313,10 @@ class LoggedMessageTest(TestCase):
             logged, [
                 LoggedMessage(logger.messages[0]),
                 LoggedMessage(logger.messages[2])])
+
+        # Lookup by string type:
+        logged2 = LoggedMessage.ofType(logger.messages, 'mymessage')
+        self.assertEqual(logged, logged2)
 
     def test_ofTypeNotFound(self):
         """